Reorganize GhcVersion
