--- src/main/scala/intellij/haskell/external/execution/GhcVersion.scala
@@ -1,14 +1,0 @@
-package intellij.haskell.external.execution
-
-case class GhcVersion(major: Int, minor: Int, patch: Int) extends Ordered[GhcVersion] {
-  def compare(that: GhcVersion): Int = GhcVersion.asc.compare(this, that)
-}
-
-object GhcVersion {
-  val asc: Ordering[GhcVersion] = Ordering.by(unapply)
-
-  def parse(version: String): GhcVersion = {
-    val parts = version.split('.')
-    GhcVersion(parts(0).toInt, parts(1).toInt, parts(2).toInt)
-  }
-}
